add button event ez430rf2500 + Resistor enable pull up/pull down (msp430 digiIO) remove WDT+ definition in models.h
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/wsim@694 c0ef0dea-404d-0410-ad65-84cceb5f200a

+++ arch/msp430/msp430_digiIO.h
@@ -52,7 +52,7 @@ void    msp430_digiIO_mcu_write         (uint16_t addr, int8_t val);
 
 int     msp430_digiIO_dev_read          (int port_number, uint8_t *val);
 uint8_t msp430_digiIO_dev_read_dir      (int port_number);
-void    msp430_digiIO_dev_write         (int port_number, uint8_t val, uint8_t bitmask);
+void    msp430_digiIO_dev_write         (int port_number, uint8_t val, uint16_t bitmask);
 
 int     msp430_digiIO_internal_dev_read (int port_number, uint8_t *val);
 void    msp430_digiIO_internal_dev_write(int port_number, uint8_t val, uint8_t bitmask);
